separate snapshot loading functions.
