+++ ed_sensor_integration/include/ed_sensor_integration/tools/snapshot.h
@@ -1,0 +1,108 @@
+#ifndef SNAPSHOT_H
+#define SNAPSHOT_H
+
+#include <ed/models/model_loader.h>
+#include <ed/world_model.h>
+#include <ed/update_request.h>
+#include <ed/serialization/serialization.h>
+#include <ed/io/json_reader.h>
+
+#include <tue/filesystem/crawler.h>
+
+#include <tue/config/read.h>
+#include <tue/config/reader.h>
+#include <tue/config/data_pointer.h>
+
+#include <rgbd/image.h>
+#include <rgbd/serialization.h>
+
+#include <vector>
+
+
+struct Snapshot
+{
+    rgbd::ImagePtr image;
+    geo::Pose3D sensor_pose;
+    std::string area_description;
+    ed::WorldModel world_model;
+};
+
+bool readImage(const std::string& filename, rgbd::ImagePtr& image, geo::Pose3D& sensor_pose,
+               ed::WorldModel& world_model, std::string& area_description)
+{
+    tue::config::DataPointer meta_data;
+
+    try
+    {
+        meta_data = tue::config::fromFile(filename);
+    }
+    catch (tue::config::ParseException& e)
+    {
+        std::cerr << "Could not open '" << filename << "'.\n\n" << e.what() << std::endl;
+        return false;
+    }
+
+    tue::config::Reader r(meta_data);
+
+    // Read image
+    std::string rgbd_filename;
+    if (r.value("rgbd_filename", rgbd_filename))
+    {
+        tue::filesystem::Path abs_rgbd_filename = tue::filesystem::Path(filename).parentPath().join(rgbd_filename);
+
+        std::ifstream f_rgbd;
+        f_rgbd.open(abs_rgbd_filename.string().c_str(), std::ifstream::binary);
+
+        if (!f_rgbd.is_open())
+        {
+            std::cerr << "Could not open '" << filename << "'." << std::endl;
+            return false;
+        }
+
+        image.reset(new rgbd::Image);
+
+        tue::serialization::InputArchive a_in(f_rgbd);
+        rgbd::deserialize(a_in, *image);
+    }
+
+    // Read sensor pose
+    if (!ed::deserialize(r, "sensor_pose", sensor_pose))
+    {
+        std::cerr << "No field 'sensor_pose' specified." << std::endl;
+        return false;
+    }
+
+//    if (r.hasError())
+//    {
+//        std::cout << "Error while reading file '" << filename << "':\n\n" << r.error() << std::endl;
+//        return false;
+//    }
+
+    return true;
+}
+
+
+bool loadWorldModel(const std::string& model_name, ed::WorldModel& world_model)
+{
+    ed::UpdateRequest req;
+
+    ed::models::ModelLoader model_loader;
+
+    std::stringstream error;
+    if (!model_loader.create("_root", model_name, req, error))
+    {
+        std::cerr << "Model '" << model_name << "' could not be loaded:" << std::endl << std::endl;
+        std::cerr << error.str() << std::endl;
+        return false;
+    }
+
+    // Reset world
+    world_model = ed::WorldModel();
+
+    // Update world
+    world_model.update(req);
+
+    return true;
+}
+
+#endif // SNAPSHOT_H
