+++ ed_sensor_integration/tools/fitter_viz.cpp
@@ -32,106 +32,13 @@
 #include <fstream>
 #include <vector>
 
-// ----------------------------------------------------------------------------------------------------
+#include "ed_sensor_integration/tools/snapshot.h"
 
-struct Snapshot
-{
-    rgbd::ImagePtr image;
-    geo::Pose3D sensor_pose;
-    std::string area_description;
-    ed::WorldModel world_model;
-};
-
-// ----------------------------------------------------------------------------------------------------
-
-bool readImage(const std::string& filename, rgbd::ImagePtr& image, geo::Pose3D& sensor_pose,
-               ed::WorldModel& world_model, std::string& area_description)
-{
-    tue::config::DataPointer meta_data;
-
-    try
-    {
-        meta_data = tue::config::fromFile(filename);
-    }
-    catch (tue::config::ParseException& e)
-    {
-        std::cerr << "Could not open '" << filename << "'.\n\n" << e.what() << std::endl;
-        return false;
-    }
-
-    tue::config::Reader r(meta_data);
-
-    // Read image
-    std::string rgbd_filename;
-    if (r.value("rgbd_filename", rgbd_filename))
-    {
-        tue::filesystem::Path abs_rgbd_filename = tue::filesystem::Path(filename).parentPath().join(rgbd_filename);
-
-        std::ifstream f_rgbd;
-        f_rgbd.open(abs_rgbd_filename.string().c_str(), std::ifstream::binary);
-
-        if (!f_rgbd.is_open())
-        {
-            std::cerr << "Could not open '" << filename << "'." << std::endl;
-            return false;
-        }
-
-        image.reset(new rgbd::Image);
-
-        tue::serialization::InputArchive a_in(f_rgbd);
-        rgbd::deserialize(a_in, *image);
-    }
-
-    // Read sensor pose
-    if (!ed::deserialize(r, "sensor_pose", sensor_pose))
-    {
-        std::cerr << "No field 'sensor_pose' specified." << std::endl;
-        return false;
-    }
-
-//    if (r.hasError())
-//    {
-//        std::cout << "Error while reading file '" << filename << "':\n\n" << r.error() << std::endl;
-//        return false;
-//    }
-
-    return true;
-}
-
-// ----------------------------------------------------------------------------------------------------
-
-bool loadWorldModel(const std::string& model_name, ed::WorldModel& world_model)
-{
-    ed::UpdateRequest req;
-
-    ed::models::ModelLoader model_loader;
-
-    std::stringstream error;
-    if (!model_loader.create("_root", model_name, req, error, true))
-    {
-        std::cerr << "Model '" << model_name << "' could not be loaded:" << std::endl << std::endl;
-        std::cerr << error.str() << std::endl;
-        return false;
-    }
-
-    // Reset world
-    world_model = ed::WorldModel();
-
-    // Update world
-    world_model.update(req);
-
-    return true;
-}
-
-// ----------------------------------------------------------------------------------------------------
 
 void usage()
 {
     std::cout << "Usage: ed_segmenter IMAGE-FILE-OR-DIRECTORY  WORLDMODEL_NAME  ENTITY_ID" << std::endl;
 }
-
-// Getting roll, pitch and yaw from a quaternion,
-// copied from https://en.wikipedia.org/wiki/Conversion_between_quaternions_and_Euler_angles
 
 void drawLine(const cv::Mat& canvas, geo::Vec2 point1, geo::Vec2 point2, geo::Transform2 pose, float resolution, int origin_x, int origin_y, cv::Scalar color)
 {
