+++ ed_sensor_integration/tools/segmenter.cpp
@@ -25,157 +25,14 @@
 #include <fstream>
 #include <vector>
 
-// ----------------------------------------------------------------------------------------------------
+#include "ed_sensor_integration/tools/snapshot.h"
 
-struct Snapshot
-{
-    rgbd::ImagePtr image;
-    geo::Pose3D sensor_pose;
-    std::string area_description;
-    ed::WorldModel world_model;
-};
-
-// ----------------------------------------------------------------------------------------------------
-
-bool readImage(const std::string& filename, rgbd::ImagePtr& image, geo::Pose3D& sensor_pose,
-               ed::WorldModel& world_model, std::string& area_description)
-{
-    tue::config::DataPointer meta_data;
-
-    try
-    {
-        meta_data = tue::config::fromFile(filename);
-    }
-    catch (tue::config::ParseException& e)
-    {
-        std::cerr << "Could not open '" << filename << "'.\n\n" << e.what() << std::endl;
-        return false;
-    }
-
-    tue::config::Reader r(meta_data);
-
-    // Read image
-    std::string rgbd_filename;
-    if (r.value("rgbd_filename", rgbd_filename))
-    {
-        tue::filesystem::Path abs_rgbd_filename = tue::filesystem::Path(filename).parentPath().join(rgbd_filename);
-
-        std::ifstream f_rgbd;
-        f_rgbd.open(abs_rgbd_filename.string().c_str(), std::ifstream::binary);
-
-        if (!f_rgbd.is_open())
-        {
-            std::cerr << "Could not open '" << filename << "'." << std::endl;
-            return false;
-        }
-
-        image.reset(new rgbd::Image);
-
-        tue::serialization::InputArchive a_in(f_rgbd);
-        rgbd::deserialize(a_in, *image);
-    }
-
-    // Read sensor pose
-    if (!ed::deserialize(r, "sensor_pose", sensor_pose))
-    {
-        std::cerr << "No field 'sensor_pose' specified." << std::endl;
-        return false;
-    }
-
-    // Reset world
-    world_model = ed::WorldModel();
-
-    // Read annotations
-    if (r.readArray("annotations"))
-    {
-        while(r.nextArrayItem())
-        {
-            std::string type;
-            double px, py;
-
-            if (!r.value("label", type) || !r.value("px", px) || !r.value("py", py))
-                continue;
-
-            // - - - - - - -
-
-            ed::UpdateRequest req;
-            ed::models::ModelLoader model_loader;
-
-            std::stringstream error;
-            ed::UUID id = "support";
-            if (model_loader.create(id, type, req, error, true))
-            {
-                // Check if this model has an 'on_top_of' volume defined
-                std::map<ed::UUID, std::map<std::string, geo::ShapeConstPtr> >::const_iterator it = req.volumes_added.find(id);
-                if (it == req.volumes_added.end())
-                    continue;
-                if (it->second.find("on_top_of") == it->second.end())
-                    continue;
-
-                int x = px * image->getDepthImage().cols;
-                int y = py * image->getDepthImage().rows;
-                rgbd::View view(*image, image->getDepthImage().cols);
-
-                geo::Vec3 pos = sensor_pose * (view.getRasterizer().project2Dto3D(x, y) * 3);
-                pos.z = 0;
-
-                req.setPose(id, geo::Pose3D(geo::Mat3::identity(), pos));
-
-                // Update world
-                world_model.update(req);
-
-                area_description = "on_top_of " + id.str();
-            }
-
-            // - - - - - - -
-
-        }
-
-        r.endArray();
-    }
-
-//    if (r.hasError())
-//    {
-//        std::cout << "Error while reading file '" << filename << "':\n\n" << r.error() << std::endl;
-//        return false;
-//    }
-
-    return true;
-}
-
-// ----------------------------------------------------------------------------------------------------
-
-bool loadWorldModel(const std::string& model_name, ed::WorldModel& world_model)
-{
-    ed::UpdateRequest req;
-
-    ed::models::ModelLoader model_loader;
-
-    std::stringstream error;
-    if (!model_loader.create("_root", model_name, req, error))
-    {
-        std::cerr << "Model '" << model_name << "' could not be loaded:" << std::endl << std::endl;
-        std::cerr << error.str() << std::endl;
-        return false;
-    }
-
-    // Reset world
-    world_model = ed::WorldModel();
-
-    // Update world
-    world_model.update(req);
-
-    return true;
-}
-
-// ----------------------------------------------------------------------------------------------------
 
 void usage()
 {
     std::cout << "Usage: ed_segmenter IMAGE-FILE-OR-DIRECTORY" << std::endl;
 }
 
-// ----------------------------------------------------------------------------------------------------
 
 int main(int argc, char **argv)
 {
